add Income & changed goals logic

+++ server/src/main/java/com/piggy/piggyServer/goals/GoalService.java
@@ -4,9 +4,11 @@ import com.piggy.piggyServer.user.UserEntity;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -15,33 +17,36 @@ public class GoalService {
 
   @Autowired
   private GoalRepository goalRepository;
-  private GoalsEntity goals;
 
-  //Crear Goal
-  public GoalsEntity createGoal(GoalsEntity goals) {
-    if (goals.getUser()== null || goals.getUser().getId() == null){
-      throw new IllegalArgumentException("User not found");
+  public GoalsEntity createGoal(GoalsEntity goal, UserEntity user){
+    if (goal.getGoalName() == null || goal.getGoalName().isEmpty()){
+      throw new IllegalArgumentException("Goal name cannot be null or empty");
     }
-    return goalRepository.save(goals);
+    if(goal.getGoalAmount() == null || goal.getGoalAmount() <= 0){
+      throw new IllegalArgumentException("Goal amount must be greater than 0");
+    }
+    goal.setUser(user);
+    return goalRepository.save(goal);
   }
 
-  //obtener todos los goals de un usuario por id
+  public ResponseEntity<?> getGoalById(Long goalId) {
+    if (!goalRepository.existsById(goalId)){
+      return ResponseEntity.status(404).body(Map.of(
+          "error", "Not found",
+          "Message", "Goal with ID" + goalId + "does not exist"
+      ));
+    }
+    GoalsEntity goal = goalRepository.findById(goalId).orElseThrow(null);
+    return ResponseEntity.ok(goal);
+  }
+
   public List<GoalsEntity> getGoalsByUserId(Long userId) {
     return goalRepository.findByUserId(userId);
   }
 
-  //obtener un goal por id
-  public GoalsEntity getGoalById(Long id) {
-    return goalRepository.findById(id).orElseThrow();
-  }
-
-  //actualizar un goal
   public GoalsEntity updateGoal(Long goalId, GoalsEntity updatedGoal) {
     GoalsEntity existingGoal = goalRepository.findById(goalId)
         .orElseThrow(() -> new RuntimeException("Goal not found"));
-    //excepción si no se encuentra el goal
-
-    //actualización de los datos
     existingGoal.setGoalName(updatedGoal.getGoalName());
     existingGoal.setSavedAmount(updatedGoal.getSavedAmount());
     existingGoal.setGoalAmount(updatedGoal.getGoalAmount());
@@ -49,12 +54,18 @@ public class GoalService {
     return goalRepository.save(existingGoal);
   }
 
-  //eliminar un goal
-  public void deleteGoalById(Long goalId) {
+  public ResponseEntity<?> deleteGoalById(Long goalId){
+    if(!goalRepository.existsById(goalId)){
+      return ResponseEntity.status(404).body(Map.of(
+          "error", "Not found",
+          "Message", "Goal with ID" + goalId + "does not exist"));
+    }
     goalRepository.deleteById(goalId);
+    return ResponseEntity.ok(Map.of(
+        "Message", "Goal with ID" + goalId + "has been successfully deleted"));
   }
 
-  //eliminar goal por nombre
+
   public void deleteGoalByName(String goalName) {
     GoalsEntity goal = goalRepository.findByGoalName(goalName);
     goalRepository.delete(goal);
